refactor: Rename ops :: utils module
Changed to common_for_install_and_uninstall
