+++ src/cargo/ops/mod.rs
@@ -41,8 +41,8 @@ mod cargo_read_manifest;
 mod cargo_run;
 mod cargo_test;
 mod cargo_uninstall;
+mod common_for_install_and_uninstall;
 mod fix;
 mod lockfile;
 mod registry;
 mod resolve;
-mod utils;
